# perf: use `Object.assign` in `@babel/plugin-proposal-object-rest-spread`